% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasight-package.R
\docType{package}
\name{seasight-package}
\alias{seasight}
\alias{seasight-package}
\title{seasight: Tools to \emph{see} seasonal adjustment more clearly}
\description{
\code{seasight} provides helper functions to run, compare and document
seasonal adjustment in a transparent and reproducible way. It builds
on the \href{https://CRAN.R-project.org/package=seasonal}{seasonal} package,
the R interface to X-13ARIMA-SEATS.
}
\details{
The main user-facing functions include:
\itemize{
\item \code{\link[=auto_seasonal_analysis]{auto_seasonal_analysis()}} for automatic model search & comparison
\item \code{\link[=sa_report_html]{sa_report_html()}} to generate HTML diagnostics reports
\item \code{\link[=sa_compare]{sa_compare()}} and \code{\link[=sa_top_candidates_table]{sa_top_candidates_table()}} for compact summaries
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://p-wegmueller.github.io/seasight/}
}

}
\author{
\strong{Maintainer}: Philipp WegmC<ller \email{philipp.wegmueller@seco.admin.ch}

}
\keyword{internal}
